requirements.txt
bcrypt==4.1.1
cachetools==5.3.2
evdev==1.6.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
keyboard==0.13.5
oauthlib==3.2.2
prompt-toolkit==3.0.36
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-xlib==0.33
questionary==2.0.1
requests-oauthlib==1.3.1
rsa==4.9
